requirements.txt
streamlit
opencv-python
pandas
Pillow
numpy
streamlit-drawable-canvas
